+++ Adjust/sdk-core/src/main/java/com/adjust/sdk/Adjust.java
@@ -32,7 +32,7 @@ public class Adjust {
      */
     public static synchronized AdjustInstance getDefaultInstance() {
         @SuppressWarnings("unused")
-        String VERSION = "!SDK-VERSION-STRING!:com.adjust.sdk:adjust-android:4.27.0";
+        String VERSION = "!SDK-VERSION-STRING!:com.adjust.sdk:adjust-android:4.28.0";
 
         if (defaultInstance == null) {
             defaultInstance = new AdjustInstance();
@@ -270,6 +270,16 @@ public class Adjust {
     public static void trackAdRevenue(final String source, final JSONObject payload) {
         AdjustInstance adjustInstance = Adjust.getDefaultInstance();
         adjustInstance.trackAdRevenue(source, payload);
+    }
+
+    /**
+     * Track ad revenue from a source provider
+     *
+     * @param adjustAdRevenue Adjust ad revenue information like source, revenue, currency etc
+     */
+    public static void trackAdRevenue(final AdjustAdRevenue adjustAdRevenue) {
+        AdjustInstance adjustInstance = Adjust.getDefaultInstance();
+        adjustInstance.trackAdRevenue(adjustAdRevenue);
     }
 
     /**