+++ Adjust/test-unit/src/androidTest/java/com/adjust/sdk/TestActivityPackage.java
@@ -53,7 +53,7 @@ public class TestActivityPackage {
         // default values
         appToken = "123456789012";
         environment = "sandbox";
-        clientSdk = "android4.27.0";
+        clientSdk = "android4.28.0";
         suffix = "";
         attribution = new AdjustAttribution();
         playServices = true;